+++ src/JsonApiBuilder.php
@@ -20,9 +20,9 @@ declare(strict_types=1);
 namespace LaravelJsonApi\Eloquent;
 
 use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\Relation as EloquentRelation;
+use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\LazyCollection;
 use Illuminate\Support\Traits\ForwardsCalls;
 use InvalidArgumentException;
@@ -41,11 +41,11 @@ use LogicException;
 use RuntimeException;
 
 /**
- * Class Builder
+ * Class JsonApiBuilder
  *
- * @mixin EloquentBuilder
+ * @mixin Builder
  */
-class Builder
+class JsonApiBuilder
 {
 
     use ForwardsCalls;
@@ -56,7 +56,7 @@ class Builder
     private Schema $schema;
 
     /**
-     * @var EloquentBuilder|EloquentRelation
+     * @var Builder|Relation
      */
     private $query;
 
@@ -81,10 +81,10 @@ class Builder
     private bool $eagerLoading = false;
 
     /**
-     * Builder constructor.
+     * JsonApiBuilder constructor.
      *
      * @param Schema $schema
-     * @param EloquentBuilder|EloquentRelation|Model $query
+     * @param Builder|Relation|Model $query
      * @param SchemaRelation|null $relation
      */
     public function __construct(Schema $schema, $query, SchemaRelation $relation = null)
@@ -93,11 +93,11 @@ class Builder
             $query = $query->newQuery();
         }
 
-        if ($query instanceof EloquentRelation && !$relation) {
+        if ($query instanceof Relation && !$relation) {
             throw new InvalidArgumentException('Expecting a schema relation when querying an Eloquent relation.');
         }
 
-        if ($relation && !$query instanceof EloquentRelation) {
+        if ($relation && !$query instanceof Relation) {
             throw new InvalidArgumentException('Expecting an Eloquent relation when querying a schema relation.');
         }
 
@@ -379,11 +379,11 @@ class Builder
     }
 
     /**
-     * @return EloquentBuilder
+     * @return Builder
      */
-    public function toBase(): EloquentBuilder
+    public function toBase(): Builder
     {
-        if ($this->query instanceof EloquentRelation) {
+        if ($this->query instanceof Relation) {
             return $this->query->getQuery();
         }